requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-jose==3.3.0
python-dotenv==1.0.0
autogluon.tabular==1.2
Pillow==10.0.0
requests==2.31.0
schedule==1.2.0 